Made some changes in line with Brad's suggestions

+++ kernel/src/platform/mpu.rs
@@ -93,7 +93,7 @@ pub trait MPU {
     fn disable_mpu(&self);
 
     /// Returns the number of supported MPU regions.
-    fn number_supported_regions(&self) -> u32;
+    fn number_supported_regions(&self) -> usize;
 
     /// Allocates a set of logical regions in the MPU.
     ///
@@ -123,7 +123,7 @@ impl MPU for () {
 
     fn disable_mpu(&self) {}
 
-    fn number_supported_regions(&self) -> u32 {
+    fn number_supported_regions(&self) -> usize {
         8
     }
